requirements.txt
torch>=1.10.0
transformers>=4.18.0
sentence-transformers>=2.2.0
spacy>=3.4.0
nltk>=3.7.0
fastapi>=0.78.0
uvicorn>=0.17.6
python-multipart>=0.0.5
regex>=2022.4.24
unidecode>=1.3.4
scikit-learn>=1.1.1
numpy>=1.22.4
pandas>=1.4.2
PyMuPDF>=1.19.6
python-docx>=0.8.11
openpyxl>=3.0.10
cryptography>=37.0.2
pytest>=7.1.2
pytest-asyncio>=0.18.3